Unshim ParquetCachedBatchSerializer
