+++ sql-plugin/src/main/scala/com/nvidia/spark/ParquetCachedBatchSerializer.scala
@@ -16,16 +16,14 @@
 
 package com.nvidia.spark
 
-import com.nvidia.spark.rapids.{DataFromReplacementRule, ExecChecks, GpuExec, RapidsConf, RapidsMeta, ShimLoader, SparkPlanMeta}
+import com.nvidia.spark.rapids.ShimLoader
 
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions.{Attribute, Expression}
 import org.apache.spark.sql.columnar.{CachedBatch, CachedBatchSerializer}
-import org.apache.spark.sql.execution.columnar.InMemoryTableScanExec
-import org.apache.spark.sql.internal.{SQLConf, StaticSQLConf}
-import org.apache.spark.sql.rapids.shims.GpuInMemoryTableScanExec
-import org.apache.spark.sql.types.{DataType, StructType}
+import org.apache.spark.sql.internal.SQLConf
+import org.apache.spark.sql.types.StructType
 import org.apache.spark.sql.vectorized.ColumnarBatch
 import org.apache.spark.storage.StorageLevel
 
@@ -38,65 +36,13 @@ trait GpuCachedBatchSerializer extends CachedBatchSerializer {
       conf: SQLConf): RDD[ColumnarBatch]
 }
 
-class InMemoryTableScanMeta(
-    imts: InMemoryTableScanExec,
-    conf: RapidsConf,
-    parent: Option[RapidsMeta[_, _, _]],
-    rule: DataFromReplacementRule)
-    extends SparkPlanMeta[InMemoryTableScanExec](imts, conf, parent, rule) {
-
-  override def tagPlanForGpu(): Unit = {
-    def stringifyTypeAttributeMap(groupedByType: Map[DataType, Set[String]]): String = {
-      groupedByType.map { case (dataType, nameSet) =>
-        dataType + " " + nameSet.mkString("[", ", ", "]")
-      }.mkString(", ")
-    }
-
-    val supportedTypeSig = rule.getChecks.get.asInstanceOf[ExecChecks]
-    val unsupportedTypes: Map[DataType, Set[String]] = imts.relation.output
-        .filterNot(attr => supportedTypeSig.check.isSupportedByPlugin(attr.dataType))
-        .groupBy(_.dataType)
-        .mapValues(_.map(_.name).toSet)
-
-    val msgFormat = "unsupported data types in output: %s"
-    if (unsupportedTypes.nonEmpty) {
-      willNotWorkOnGpu(msgFormat.format(stringifyTypeAttributeMap(unsupportedTypes)))
-    }
-    if (!imts.relation.cacheBuilder.serializer
-        .isInstanceOf[com.nvidia.spark.ParquetCachedBatchSerializer]) {
-      willNotWorkOnGpu("ParquetCachedBatchSerializer is not being used")
-      if (SQLConf.get.getConf(StaticSQLConf.SPARK_CACHE_SERIALIZER)
-          .equals("com.nvidia.spark.ParquetCachedBatchSerializer")) {
-        throw new IllegalStateException("Cache serializer failed to load! " +
-            "Something went wrong while loading ParquetCachedBatchSerializer class")
-      }
-    }
-  }
-  /**
-   * Convert InMemoryTableScanExec to a GPU enabled version.
-   */
-  override def convertToGpu(): GpuExec = {
-    GpuInMemoryTableScanExec(imts.attributes, imts.predicates, imts.relation)
-  }
-}
-
 /**
- * User facing wrapper class that calls into the proper shim version.
+ * User facing wrapper class that calls into the internal version.
  */
 class ParquetCachedBatchSerializer extends GpuCachedBatchSerializer {
 
-  val minSupportedVer = "3.1.0"
-  val sparkVersion = ShimLoader.getSparkVersion
-  // Note that since the config to set the serializer wasn't added until
-  // Spark 3.1.0 (https://issues.apache.org/jira/browse/SPARK-32274) this shouldn't
-  // ever throw.
-  if (sparkVersion < minSupportedVer) {
-    throw new IllegalArgumentException("ParquetCachedBaatchSerializer only supported for Spark " +
-      s"versions > 3.1.0, version found was: $sparkVersion")
-  }
-
   private lazy val realSerializer: GpuCachedBatchSerializer = {
-    ShimLoader.newInstanceOf("com.nvidia.spark.rapids.shims.ParquetCachedBatchSerializer")
+    ShimLoader.newInstanceOf("com.nvidia.spark.rapids.ParquetCachedBatchSerializer")
   }
 
   /**
